Refactored the page header component. It is now a directory with a parent component which will instantiate the correct child component depending on the current background color
